+++ CHANGELOG.md
@@ -7,7 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-*Nothing yet.*
+### Fixed
+
+* Fix error message when an invalid access token is provided.
+* Fixed imports not starting in 1.0.0-rc.2.
 
 ## [1.0.0-rc.2] - January 3, 2020